refactor(ui): extract sidebar into dbflux_ui_sidebar crate (#102 Phase 5)
Extracts the views/sidebar/ subtree (14 files, ~16.5k LOC: the connection and
scripts sidebar tree) out of dbflux_ui into a new dbflux_ui_sidebar crate. The
seam is one-way (workspace depends on sidebar, sidebar has zero references to
workspace), so the new crate sits cleanly below the workspace integrator with
no cycle. This is the single largest remaining chunk; dbflux_ui drops to ~11.5k.

Prerequisite down-move: sql_preview_modal.rs (sidebar's only upward edge, also
used by workspace) -> dbflux_ui_base, mirroring sso_wizard/platform; a shim is
left at the old path so workspace keeps resolving SqlPreviewModal.

Behavior-preserving relocation: imports rewritten to the downward crates
(dbflux_components / dbflux_core / dbflux_app / dbflux_ui_base / dbflux_ui_windows),
feature passthroughs (mcp/lua/aws) replicated, and a pub-use shim at
dbflux_ui/src/ui/views/sidebar/mod.rs keeps every crate::ui::views::sidebar::*
call-site in workspace and dock unchanged. Three sidebar operations methods were
widened pub(crate) -> pub for cross-crate use by workspace; no semantics changed.
Test parity held at 2322.

Part of #102 (Phase 5 of the multi-crate split).
