+++ crates/dbflux_ui/Cargo.toml
@@ -17,6 +17,7 @@ dbflux_components.workspace = true
 dbflux_ui_base.workspace = true
 dbflux_ui_document.workspace = true
 dbflux_ui_windows.workspace = true
+dbflux_ui_sidebar.workspace = true
 dbflux_core.workspace = true
 dbflux_storage.workspace = true
 dbflux_app.workspace = true
@@ -100,9 +101,9 @@ redis = ["dbflux_driver_redis", "dbflux_mcp_server?/redis"]
 dynamodb = ["dbflux_driver_dynamodb", "dbflux_mcp_server?/dynamodb"]
 cloudwatch = ["dbflux_driver_cloudwatch"]
 mssql = ["dbflux_driver_mssql"]
-lua = ["dbflux_lua", "dbflux_app/lua", "dbflux_ui_base/lua", "dbflux_ui_document/lua", "dbflux_ui_windows/lua"]
-aws = ["dbflux_aws", "dbflux_ssm", "dbflux_mcp_server?/aws", "dbflux_app/aws", "dbflux_ui_base/aws", "dbflux_ui_document/aws", "dbflux_ui_windows/aws"]
-mcp = ["dbflux_mcp", "dbflux_mcp_server", "dbflux_policy", "dbflux_approval", "dbflux_app/mcp", "dbflux_ui_base/mcp", "dbflux_ui_document/mcp", "dbflux_ui_windows/mcp"]
+lua = ["dbflux_lua", "dbflux_app/lua", "dbflux_ui_base/lua", "dbflux_ui_document/lua", "dbflux_ui_windows/lua", "dbflux_ui_sidebar/lua"]
+aws = ["dbflux_aws", "dbflux_ssm", "dbflux_mcp_server?/aws", "dbflux_app/aws", "dbflux_ui_base/aws", "dbflux_ui_document/aws", "dbflux_ui_windows/aws", "dbflux_ui_sidebar/aws"]
+mcp = ["dbflux_mcp", "dbflux_mcp_server", "dbflux_policy", "dbflux_approval", "dbflux_app/mcp", "dbflux_ui_base/mcp", "dbflux_ui_document/mcp", "dbflux_ui_windows/mcp", "dbflux_ui_sidebar/mcp"]
 vendored-openssl = ["dbflux_ssh/vendored-openssl"]
 
 [dev-dependencies]
