+++ crates/dbflux_ui/src/ui/mod.rs
@@ -11,7 +11,3 @@ pub mod windows;
 
 #[cfg(test)]
 mod design_system_guardrails;
-
-// AsyncUpdateResultExt now lives in dbflux_ui_base; re-export at this path so
-// all ~9 existing `use crate::ui::AsyncUpdateResultExt` call-sites are unchanged.
-pub(crate) use dbflux_ui_base::AsyncUpdateResultExt;
