+++ crates/dbflux_ui_sidebar/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "dbflux_ui_sidebar"
+edition.workspace = true
+version.workspace = true
+authors.workspace = true
+description = "Connection and scripts sidebar tree for DBFlux"
+repository.workspace = true
+homepage.workspace = true
+license.workspace = true
+
+[lib]
+
+[dependencies]
+gpui.workspace = true
+gpui-component.workspace = true
+dbflux_ui_base.workspace = true
+dbflux_ui_windows.workspace = true
+dbflux_components.workspace = true
+dbflux_core.workspace = true
+dbflux_app.workspace = true
+dbflux_ssh.workspace = true
+uuid.workspace = true
+log.workspace = true
+serde_json.workspace = true
+rfd.workspace = true
+dirs.workspace = true
+anyhow = "1"
+
+[features]
+default = []
+# Passthroughs: sidebar code has no direct mcp/lua/aws gates, but these must
+# forward to dbflux_app/dbflux_ui_base so dbflux_app is not compiled twice
+# with mismatched feature sets across the workspace.
+mcp = ["dbflux_app/mcp", "dbflux_ui_base/mcp"]
+lua = ["dbflux_app/lua", "dbflux_ui_base/lua"]
+aws = ["dbflux_app/aws", "dbflux_ui_base/aws"]
