+++ crates/dbflux_ui_sidebar/src/context_menu.rs
