+++ crates/dbflux_ui_sidebar/src/deletion.rs
