+++ crates/dbflux_ui_sidebar/src/expansion.rs
