+++ crates/dbflux_ui_sidebar/src/render.rs
@@ -338,7 +338,7 @@ impl Sidebar {
 
 impl Render for Sidebar {
     fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        crate::ui::components::toast::flush_pending_toast(self.pending_toast.take(), window, cx);
+        dbflux_ui_base::toast::flush_pending_toast(self.pending_toast.take(), window, cx);
 
         if let Some(item_id) = self.pending_rename_item.take() {
             self.start_rename(&item_id, window, cx);
