+++ crates/dbflux_ui_sidebar/src/selection.rs
