+++ crates/dbflux_ui_sidebar/src/tree_builder.rs
